seperated functionality into OO-based webrepl API (webrepl_client) that can be used by other python programs, and refactored webrepl_cli to use this library
